refactor: pathutil
